refactor: extract event bus into standalone _event_bus/ module
Move server, state, protocol from _display/socket/ into _event_bus/.
SocketDisplay and client become thin consumers. Any part of Inspect
can now emit/consume events without depending on the display layer.
