+++ src/inspect_ai/_event_bus/__init__.py
@@ -1,0 +1,25 @@
+from .protocol import (
+    CancelSampleCommand,
+    ClientMessage,
+    EvalCompleteMessage,
+    MetricValue,
+    MetricsUpdateMessage,
+    PrintMessage,
+    ProgressUpdateMessage,
+    SampleCancelledMessage,
+    SampleCompleteMessage,
+    SampleEndMessage,
+    SampleInfo,
+    SampleStartMessage,
+    ServerMessage,
+    SnapshotMessage,
+    TaskCompleteMessage,
+    TaskInfo,
+    TaskProgress,
+    TaskStartMessage,
+    parse_client_message,
+    parse_server_message,
+    to_json_line,
+)
+from .server import SocketServer
+from .state import StateManager
